# Optimize `copy_cells`, part 1